🚚 move soft remove to delete.rs
